[FIX] 순환참조로인한 OrderPointService 분리
