+++ src/main/java/com/programmers/smrtstore/domain/orderManagement/order/application/OrderServiceImpl.java
@@ -146,30 +146,6 @@ public class OrderServiceImpl implements OrderService {
     }
 
     @Override
-    public Integer calculateUserMonthlyTotalSpending(Long userId, int month, int year) {
-        checkUserExistence(userId);
-        DateTimeUtils.validateMonth(month);
-        DateTimeUtils.validateYear(year);
-        return orderJpaRepository.calculateMonthlyTotalSpending(userId, month, year);
-    }
-
-    @Override
-    public Integer getTotalPriceByOrderId(String orderId) {
-        Order order = orderJpaRepository.findByIdIncludeDeleted(orderId)
-            .orElseThrow(() -> new OrderException(ORDER_NOT_FOUND, String.valueOf(orderId)));
-        return order.getTotalPrice();
-    }
-
-    @Override
-    public List<OrderedProductResponse> getProductsForOrder(String orderId) {
-        return orderJpaRepository.findByIdWithOrderSheetIncludeDeleted(orderId)
-            .orElseThrow(() -> new OrderException(ORDER_NOT_FOUND, String.valueOf(orderId)))
-            .getOrderSheet().getOrderedProducts().stream()
-            .map(OrderedProductResponse::from)
-            .toList();
-    }
-
-    @Override
     public List<OrderPreviewResponse> getOrderPreviewsByUserId(Long userId) {
         checkUserExistence(userId);
         List<Order> orders = orderJpaRepository.findByUserId(userId);
