+++ src/main/java/com/programmers/smrtstore/domain/point/application/PointService.java
@@ -1,7 +1,7 @@
 package com.programmers.smrtstore.domain.point.application;
 
 import com.programmers.smrtstore.core.properties.ErrorCode;
-import com.programmers.smrtstore.domain.orderManagement.order.application.OrderService;
+import com.programmers.smrtstore.domain.orderManagement.order.application.OrderPointService;
 import com.programmers.smrtstore.domain.orderManagement.order.presentation.dto.res.OrderedProductResponse;
 import com.programmers.smrtstore.domain.point.application.dto.req.PointRequest;
 import com.programmers.smrtstore.domain.point.application.dto.req.UsePointRequest;
@@ -37,7 +37,7 @@ import org.springframework.transaction.annotation.Transactional;
 public class PointService {
 
     private final PointFacade pointFacade;
-    private final OrderService orderService;
+    private final OrderPointService orderPointService;
     private final ReviewJpaRepository reviewRepository;
     private final ProductJpaRepository productRepository;
     private final UserJpaRepository userJpaRepository;
@@ -64,7 +64,7 @@ public class PointService {
 
         // 주문상품 리스트와 상품별 결제가격 조회
         int totalPoint = 0;
-        List<OrderedProductResponse> orderedProducts = orderService.getProductsForOrder(orderId);
+        List<OrderedProductResponse> orderedProducts = orderPointService.getProductsForOrder(orderId);
         for (OrderedProductResponse product : orderedProducts) {
             // 상품별 (1개당) 실제 적립되는 적립금 (기본 1% 적립 + 추가 4% 적립)
             int totalAcmPoint = calculateAcmPointPerProduct(product, userMonthlyTotalSpending);
@@ -125,7 +125,7 @@ public class PointService {
 
         int additionalPoint = 0;
         if (membershipYn) {
-            List<OrderedProductResponse> orderedProducts = orderService.getProductsForOrder(orderId);
+            List<OrderedProductResponse> orderedProducts = orderPointService.getProductsForOrder(orderId);
             // 멤버십, 월별 쇼핑 금액이 반영된 추가 멤버십 적용 금액
             additionalPoint = calculateAdditionalAcmPoint(orderedProducts, userId);
         }
@@ -158,7 +158,7 @@ public class PointService {
     }
 
     private int calculateDefaultPoint(String orderId) {
-        return orderService.getTotalPriceByOrderId(orderId) / 100;
+        return orderPointService.getTotalPriceByOrderId(orderId) / 100;
     }
 
     public ProductEstimatedPointDto calculateEstimatedAcmPoint(Long productId, Long userId) {
@@ -234,7 +234,7 @@ public class PointService {
         int year = now.getYear();
         int month = now.getMonthValue();
 
-        return orderService.calculateUserMonthlyTotalSpending(userId, month, year);
+        return orderPointService.calculateUserMonthlyTotalSpending(userId, month, year);
     }
 
     private int calculateAdditionalPoint(List<OrderedProductResponse> orderedProducts, int userMonthlyTotalSpending) {
